Refactoring Image (partial)
* `Image` is particularly challenging to refactor because it's entangled
with other classes.
* Shoes3/shoes3#333
