+++ shoes/canvas.h
@@ -306,47 +306,12 @@ VALUE shoes_canvas_set_scroll_top(VALUE, VALUE);
 VALUE shoes_canvas_get_gutter_width(VALUE);
 VALUE shoes_canvas_displace(VALUE, VALUE, VALUE);
 VALUE shoes_canvas_move(VALUE, VALUE, VALUE);
-VALUE shoes_canvas_nostroke(VALUE);
-VALUE shoes_canvas_stroke(int, VALUE *, VALUE);
-VALUE shoes_canvas_strokewidth(VALUE, VALUE);
-VALUE shoes_canvas_dash(VALUE, VALUE);
-VALUE shoes_canvas_cap(VALUE, VALUE);
-VALUE shoes_canvas_nofill(VALUE);
-VALUE shoes_canvas_fill(int, VALUE *, VALUE);
 VALUE shoes_canvas_rgb(int, VALUE *, VALUE);
 VALUE shoes_canvas_gray(int, VALUE *, VALUE);
-VALUE shoes_canvas_rect(int, VALUE *, VALUE);
-VALUE shoes_canvas_arc(int, VALUE *, VALUE);
-VALUE shoes_canvas_oval(int, VALUE *, VALUE);
-VALUE shoes_canvas_line(int, VALUE *, VALUE);
-VALUE shoes_canvas_arrow(int, VALUE *, VALUE);
-VALUE shoes_canvas_star(int, VALUE *, VALUE);
-VALUE shoes_canvas_para(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_banner(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_title(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_subtitle(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_tagline(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_caption(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_inscription(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_background(int, VALUE *, VALUE);
-VALUE shoes_canvas_border(int, VALUE *, VALUE);
-VALUE shoes_canvas_blur(int, VALUE *, VALUE);
-VALUE shoes_canvas_glow(int, VALUE *, VALUE);
-VALUE shoes_canvas_shadow(int, VALUE *, VALUE);
-VALUE shoes_canvas_image(int, VALUE *, VALUE);
 VALUE shoes_canvas_plot(int, VALUE *, VALUE);
 VALUE shoes_canvas_chart_series(int, VALUE *, VALUE);
 VALUE shoes_canvas_imagesize(VALUE, VALUE);
 void shoes_canvas_remove_item(VALUE, VALUE, char, char);
-VALUE shoes_canvas_move_to(VALUE, VALUE, VALUE);
-VALUE shoes_canvas_line_to(VALUE, VALUE, VALUE);
-VALUE shoes_canvas_curve_to(VALUE, VALUE, VALUE, VALUE, VALUE, VALUE, VALUE);
-VALUE shoes_canvas_arc_to(VALUE, VALUE, VALUE, VALUE, VALUE, VALUE, VALUE);
-VALUE shoes_canvas_transform(VALUE, VALUE);
-VALUE shoes_canvas_translate(VALUE, VALUE, VALUE);
-VALUE shoes_canvas_rotate(VALUE, VALUE);
-VALUE shoes_canvas_scale(int, VALUE *, VALUE);
-VALUE shoes_canvas_skew(int, VALUE *, VALUE);
 VALUE shoes_canvas_push(VALUE);
 VALUE shoes_canvas_pop(VALUE);
 VALUE shoes_canvas_reset(VALUE);
