+++ shoes/types/pattern.h
@@ -12,12 +12,6 @@
 extern VALUE cShoes, cApp, cTypes, cCanvas, cWidget;
 extern shoes_app _shoes_app;
 
-// native forward declarations
-// extern SHOES_CONTROL_REF shoes_native_spinner(VALUE self, shoes_canvas *canvas, shoes_place *place, VALUE attr, char *msg);
-// extern void shoes_native_spinner_start(SHOES_CONTROL_REF ref);
-// extern void shoes_native_spinner_stop(SHOES_CONTROL_REF ref);
-// extern gboolean shoes_native_spinner_started(SHOES_CONTROL_REF ref);
-
 VALUE cPattern, cBorder, cBackground;
 
 typedef struct {
@@ -47,6 +41,10 @@ VALUE shoes_pattern_alloc(VALUE klass);
 VALUE shoes_background_draw(VALUE self, VALUE c, VALUE actual);
 VALUE shoes_border_draw(VALUE self, VALUE c, VALUE actual);
 VALUE shoes_subpattern_new(VALUE klass, VALUE pat, VALUE parent);
+
+// canvas
+VALUE shoes_canvas_background(int argc, VALUE *argv, VALUE self);
+VALUE shoes_canvas_border(int argc, VALUE *argv, VALUE self);
 
 #define PATH_OUT(cr, attr, place, sw, cap, dash, pen, cfunc) \
 { \
@@ -119,10 +117,5 @@ VALUE shoes_subpattern_new(VALUE klass, VALUE pat, VALUE parent);
   { \
     cairo_pattern_set_matrix(PATTERN(self_t), &matrix1); \
   }
-  
-
-// canvas
-VALUE shoes_canvas_background(int argc, VALUE *argv, VALUE self);
-VALUE shoes_canvas_border(int argc, VALUE *argv, VALUE self);
 
 #endif
